requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pillow>=10.0.0
